requirements.txt
azure-core==1.39.0
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pluggy==1.6.0
py4j==0.10.9.9
pycparser==3.0
Pygments==2.20.0
pyspark==4.1.1
pytest==9.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
